requirement.txt
gradio
azure-cognitiveservices-speech